rename files and components for consistency, add barrel files for cleaner import
