+++ packages/components/src/Combobox/components/ComboboxContent/ComboboxContentList/index.ts
@@ -1,0 +1,1 @@
+export { ComboboxContentList } from "./ComboboxContentList";
