+++ packages/components/src/Combobox/components/ComboboxContent/ComboboxContentSearch/index.ts
@@ -1,0 +1,1 @@
+export { ComboboxContentSearch } from "./ComboboxContentSearch";
